requirements.txt
llama-parse
python-dotenv
docx
pypdf
csv
tempfile